+++ src/createTransformer.ts
@@ -21,15 +21,17 @@ import { minifyTemplate } from './minify';
  * styled(Component)
  * styled('tag')
  * styledFunction.attrs(attributes)
-*/
-function isStyledFunction(node: ts.Node, identifiers: CustomStyledIdentifiers): boolean {
+ */
+function isStyledFunction(
+    node: ts.Node,
+    identifiers: CustomStyledIdentifiers
+): node is ts.PropertyAccessExpression | ts.CallExpression {
     if (isPropertyAccessExpression(node)) {
         if (isStyledObject(node.expression, identifiers)) {
             return true;
         }
 
-        if (isStyledExtendIdentifier(node.name.text, identifiers)
-            && isValidComponent(node.expression)) {
+        if (isStyledExtendIdentifier(node.name.text, identifiers) && isValidComponent(node.expression)) {
             return true;
         }
 
@@ -37,7 +39,6 @@ function isStyledFunction(node: ts.Node, identifiers: CustomStyledIdentifiers): 
     }
 
     if (isCallExpression(node) && node.arguments.length === 1) {
-
         if (isStyledObject(node.expression, identifiers)) {
             return true;
         }
@@ -71,7 +72,7 @@ function isValidTagName(name: string) {
 }
 
 function isValidComponentName(name: string) {
-    return isLetter(name[0]) && (name[0] === name[0].toUpperCase());
+    return isLetter(name[0]) && name[0] === name[0].toUpperCase();
 }
 
 function isStyledAttrsIdentifier(name: string, { attrs: attrsIdentifiers = ['attrs'] }: CustomStyledIdentifiers) {
@@ -79,9 +80,12 @@ function isStyledAttrsIdentifier(name: string, { attrs: attrsIdentifiers = ['att
 }
 
 function isStyledAttrs(node: ts.Node, identifiers: CustomStyledIdentifiers) {
-    return node && isPropertyAccessExpression(node)
-        && isStyledAttrsIdentifier(node.name.text, identifiers)
-        && isStyledFunction((node as ts.PropertyAccessExpression).expression, identifiers);
+    return (
+        node &&
+        isPropertyAccessExpression(node) &&
+        isStyledAttrsIdentifier(node.name.text, identifiers) &&
+        isStyledFunction((node as ts.PropertyAccessExpression).expression, identifiers)
+    );
 }
 
 function isStyledKeyframesIdentifier(name: string, { keyframes = ['keyframes'] }: CustomStyledIdentifiers) {
@@ -92,7 +96,10 @@ function isStyledCssIdentifier(name: string, { css = ['css'] }: CustomStyledIden
     return css.indexOf(name) >= 0;
 }
 
-function isStyledCreateGlobalStyleIdentifier(name: string, { createGlobalStyle = ['createGlobalStyle'] }: CustomStyledIdentifiers) {
+function isStyledCreateGlobalStyleIdentifier(
+    name: string,
+    { createGlobalStyle = ['createGlobalStyle'] }: CustomStyledIdentifiers
+) {
     return createGlobalStyle.indexOf(name) >= 0;
 }
 
@@ -101,55 +108,59 @@ function isStyledExtendIdentifier(name: string, { extend = [] }: CustomStyledIde
 }
 
 function isMinifyableStyledFunction(node: ts.Node, identifiers: CustomStyledIdentifiers) {
-    return isStyledFunction(node, identifiers)
-        || (
-            isIdentifier(node)
-            && (
-                isStyledKeyframesIdentifier(node.text, identifiers)
-                || isStyledCssIdentifier(node.text, identifiers)
-                || isStyledCreateGlobalStyleIdentifier(node.text, identifiers)
-            )
-        );
+    return (
+        isStyledFunction(node, identifiers) ||
+        (isIdentifier(node) &&
+            (isStyledKeyframesIdentifier(node.text, identifiers) ||
+                isStyledCssIdentifier(node.text, identifiers) ||
+                isStyledCreateGlobalStyleIdentifier(node.text, identifiers)))
+    );
 }
 
 function defaultGetDisplayName(filename: string, bindingName: string | undefined): string | undefined {
     return bindingName;
 }
 
-export function createTransformer(options?: Partial<Options>): ts.TransformerFactory<ts.SourceFile>
+export function createTransformer(options?: Partial<Options>): ts.TransformerFactory<ts.SourceFile>;
 export function createTransformer({
     getDisplayName = defaultGetDisplayName,
     identifiers = {},
     ssr = true,
     displayName = true,
     minify = false,
-    componentIdPrefix = '',
-} : Partial<Options> = {}) {
-
+    componentIdPrefix = ''
+}: Partial<Options> = {}) {
     /**
      * Infers display name of a styled component.
      * Recognizes the following patterns:
      *
      * (const|var|let) ComponentName = styled...
      * export default styled...
-    */
-    function getDisplayNameFromNode(node: ts.Node, componentIdPrefix?: string): string | undefined {
+     */
+    function getDisplayNameFromNode(node: ts.Node, sourceFile: ts.SourceFile): string | undefined {
         if (isVariableDeclaration(node) && isIdentifier(node.name)) {
-            return (!!componentIdPrefix ? `${componentIdPrefix}-` : '') + getDisplayName(node.getSourceFile().fileName, node.name.text);
+            return (componentIdPrefix ? componentIdPrefix + '-' : '') + getDisplayName(sourceFile.fileName, node.name.text);
         }
 
         if (isExportAssignment(node)) {
-            return getDisplayName(node.getSourceFile().fileName, undefined);
+            return getDisplayName(sourceFile.fileName, undefined);
         }
 
         return undefined;
     }
 
-    function getIdFromNode(node: ts.Node, sourceRoot: string | undefined, position: number, componentIdPrefix?: string): string | undefined {
+    function getIdFromNode(
+        node: ts.Node,
+        sourceRoot: string | undefined,
+        position: number,
+        sourceFile: ts.SourceFile
+    ): string | undefined {
         if ((isVariableDeclaration(node) && isIdentifier(node.name)) || isExportAssignment(node)) {
-            const fileName = node.getSourceFile().fileName;
-            const filePath = sourceRoot ? path.relative(sourceRoot, fileName).replace(path.sep, path.posix.sep) : fileName;
-            return `${!!componentIdPrefix ? componentIdPrefix : 'sc'}-` + hash(`${getDisplayNameFromNode(node)}${filePath}${position}`);
+            const fileName = sourceFile.fileName;
+            const filePath = sourceRoot
+                ? path.relative(sourceRoot, fileName).replace(path.sep, path.posix.sep)
+                : fileName;
+            return (componentIdPrefix ?? 'sc') + '-' + hash(`${getDisplayNameFromNode(node, sourceFile)}${filePath}${position}`);
         }
         return undefined;
     }
@@ -157,67 +168,121 @@ export function createTransformer({
     const transformer: ts.TransformerFactory<ts.SourceFile> = (context) => {
         const { sourceRoot } = context.getCompilerOptions();
 
-        return (node) => {
+        return (sourceFile) => {
             let lastComponentPosition = 0;
 
-            const visitor: ts.Visitor = (node) => {
-                if (
-                    minify
-                    && isTaggedTemplateExpression(node)
-                    && isMinifyableStyledFunction(node.tag, identifiers)
-                ) {
-                    const minifiedTemplate = minifyTemplate(node.template);
-                    if (minifiedTemplate && minifiedTemplate !== node.template) {
-                        const newNode = ts.createTaggedTemplate(node.tag, node.typeArguments, minifiedTemplate);
-                        newNode.parent = node.parent;
-                        node = newNode;
-                    }
+            const withConfig = (node: ts.Expression, properties: ts.PropertyAssignment[]) =>
+                properties.length > 0
+                    ? context.factory.createCallExpression(
+                          context.factory.createPropertyAccessExpression(node, 'withConfig'),
+                          undefined,
+                          [context.factory.createObjectLiteralExpression(properties)]
+                      )
+                    : node;
+
+            const createDisplayNameConfig = (displayNameValue: string | undefined) =>
+                displayNameValue
+                    ? [
+                          context.factory.createPropertyAssignment(
+                              'displayName',
+                              context.factory.createStringLiteral(displayNameValue)
+                          ),
+                      ]
+                    : [];
+            const createIdConfig = (componentId: string | undefined) =>
+                componentId
+                    ? [
+                          context.factory.createPropertyAssignment(
+                              'componentId',
+                              context.factory.createStringLiteral(componentId)
+                          ),
+                      ]
+                    : [];
+
+            const transformStyledFunction = (
+                binding: ts.VariableDeclaration | ts.ExportAssignment,
+                node: ts.Expression
+            ) =>
+                withConfig(node, [
+                    ...(displayName ? createDisplayNameConfig(getDisplayNameFromNode(binding, sourceFile)) : []),
+                    ...(ssr
+                        ? createIdConfig(getIdFromNode(binding, sourceRoot, ++lastComponentPosition, sourceFile))
+                        : []),
+                ]);
+
+            const transformTemplateLiteral = (templateLiteral: ts.TemplateLiteral) =>
+                minify ? minifyTemplate(templateLiteral, context.factory) : templateLiteral;
+
+            const transformTaggedTemplateExpression = (node: ts.TaggedTemplateExpression) =>
+                isMinifyableStyledFunction(node.tag, identifiers)
+                    ? context.factory.updateTaggedTemplateExpression(
+                          node,
+                          node.tag,
+                          node.typeArguments,
+                          transformTemplateLiteral(node.template)
+                      )
+                    : node;
+
+            const transformBindingExpression = (
+                binding: ts.VariableDeclaration | ts.ExportAssignment,
+                node: ts.Expression
+            ) => {
+                if (isTaggedTemplateExpression(node) && isStyledFunction(node.tag, identifiers)) {
+                    return context.factory.updateTaggedTemplateExpression(
+                        node,
+                        transformStyledFunction(binding, node.tag),
+                        node.typeArguments,
+                        transformTemplateLiteral(node.template)
+                    );
                 }
-
-                if (
-                    node.parent
-                    && (
-                        isTaggedTemplateExpression(node.parent) && node.parent.tag === node
-                        || isCallExpression(node.parent)
-                    )
-                    && node.parent.parent
-                    && (isVariableDeclaration(node.parent.parent) || isExportAssignment(node.parent.parent))
-                    && isStyledFunction(node, identifiers)
-                ) {
-
-                    const styledConfig = [];
-
-                    if (displayName) {
-                        const displayNameValue = getDisplayNameFromNode(node.parent.parent, componentIdPrefix);
-                        if (displayNameValue) {
-                            styledConfig.push(ts.createPropertyAssignment('displayName', ts.createLiteral(displayNameValue)));
-                        }
-                    }
-
-                    if (ssr) {
-                        const componentId = getIdFromNode(node.parent.parent, sourceRoot, ++lastComponentPosition, componentIdPrefix);
-                        if (componentId) {
-                            styledConfig.push(ts.createPropertyAssignment('componentId', ts.createLiteral(componentId)));
-                        }
-                    }
-
-                    if (styledConfig.length > 0) {
-                        return ts.createCall(
-                            ts.createPropertyAccess(node as ts.Expression, 'withConfig'),
-                            undefined,
-                            [ts.createObjectLiteral(styledConfig)]);
-                    }
+                if (isCallExpression(node) && isStyledFunction(node.expression, identifiers)) {
+                    return context.factory.updateCallExpression(
+                        node,
+                        transformStyledFunction(binding, node.expression),
+                        node.typeArguments,
+                        node.arguments
+                    );
                 }
+            };
 
-                ts.forEachChild(node, n => {
-                    if (!n.parent)
-                        n.parent = node;
-                });
+            const updateNode = <T extends ts.Node, D>(
+                node: T,
+                data: D | undefined,
+                updateFn: (node: T, data: D) => T
+            ) => (data ? updateFn(node, data) : undefined);
 
-                return ts.visitEachChild(node, visitor, context);
-            }
+            const updateVariableDeclarationInitializer = (node: ts.VariableDeclaration, initializer: ts.Expression) =>
+                context.factory.updateVariableDeclaration(
+                    node,
+                    node.name,
+                    node.exclamationToken,
+                    node.type,
+                    initializer
+                );
 
-            return ts.visitNode(node, visitor);
+            const updateExportAssignmentExpression = (node: ts.ExportAssignment, expression: ts.Expression) =>
+                context.factory.updateExportAssignment(node, node.decorators, node.modifiers, expression);
+
+            const transformNode = (node: ts.Node) =>
+                isVariableDeclaration(node) && node.initializer
+                    ? updateNode(
+                          node,
+                          transformBindingExpression(node, node.initializer),
+                          updateVariableDeclarationInitializer
+                      )
+                    : isExportAssignment(node)
+                    ? updateNode(
+                          node,
+                          transformBindingExpression(node, node.expression),
+                          updateExportAssignmentExpression
+                      )
+                    : minify && isTaggedTemplateExpression(node)
+                    ? transformTaggedTemplateExpression(node)
+                    : undefined;
+
+            const visitNode: ts.Visitor = (node) => transformNode(node) || ts.visitEachChild(node, visitNode, context);
+
+            return ts.visitNode(sourceFile, visitNode);
         };
     };